requirements.txt
Django==2.1.3
djangorestframework==3.9.0
flake8==3.9.1
psycopg2-binary==2.7.5